add flag for launch tests

+++ verifier-guest.sh
@@ -72,20 +72,26 @@ sudo chown -R www-data.www-data /var/www/html
 
 # deleted index.html from /var/www/html
 sudo rm /var/www/html/index.html
-sleep 40000
+# sleep 40000
 
 #install selenium,pip
-cd ~ 
-sudo apt-get -y install python-pip
-sudo pip install --upgrade pip
-sudo pip install nose
-sudo pip install -U selenium==3.0.1
-wget http://ftp.openquake.org/mirror/mozilla/geckodriver-latest-linux64.tar.gz ; tar zxvf geckodriver-latest-linux64.tar.gz ; sudo cp geckodriver /usr/local/bin
+cd ~
 
-cp $GEM_GIT_PACKAGE/openquake/taxonomy/test/config/moon_config.py.tmpl $GEM_GIT_PACKAGE/openquake/taxonomy/test/config/moon_config.py
-git clone -b "$branch_id" --depth=1  $GEM_GIT_REPO/oq-moon.git || git clone --depth=1 $GEM_GIT_REPO/oq-moon.git
+# if variable not valorized don't begin tests
+NO_EXEC_TEST=$3
 
-export DISPLAY=:1
-export PYTHONPATH=oq-moon:$GEM_GIT_PACKAGE:$GEM_GIT_PACKAGE/openquake/taxonomy/test/config
-python -m openquake.moon.nose_runner --failurecatcher prod -s -v --with-xunit --xunit-file=xunit-platform-prod.xml $GEM_GIT_PACKAGE/openquake/taxonomy/test || true
-# sleep 40000 || true
+if [ -f  NO_EXEC_TEST != "notest" ] ; then 
+    sudo apt-get -y install python-pip
+    sudo pip install --upgrade pip
+    sudo pip install nose
+    sudo pip install -U selenium==3.0.1
+    wget http://ftp.openquake.org/mirror/mozilla/geckodriver-latest-linux64.tar.gz ; tar zxvf geckodriver-latest-linux64.tar.gz ; sudo cp geckodriver /usr/local/bin
+
+    cp $GEM_GIT_PACKAGE/openquake/taxonomy/test/config/moon_config.py.tmpl $GEM_GIT_PACKAGE/openquake/taxonomy/test/config/moon_config.py
+    git clone -b "$branch_id" --depth=1  $GEM_GIT_REPO/oq-moon.git || git clone --depth=1 $GEM_GIT_REPO/oq-moon.git
+
+    export DISPLAY=:1
+    export PYTHONPATH=oq-moon:$GEM_GIT_PACKAGE:$GEM_GIT_PACKAGE/openquake/taxonomy/test/config
+    python -m openquake.moon.nose_runner --failurecatcher prod -s -v --with-xunit --xunit-file=xunit-platform-prod.xml $GEM_GIT_PACKAGE/openquake/taxonomy/test || true
+    # sleep 40000 || true
+fi
